label and property fixes for formatSuggestions
label handelling in details.jsx was a mess since format suggestions was sending outlaws to details.jsx This issue is temporarily fixed. To look up flight numbr just type 4 it will drop down flight numbers, select any and it should show livedata for that particular flight number. also these work on searchbar raw submits - 414 or ua414 or UA4433 etc. It's been handeled in backend and is only a temporary fix