NAS cleanup, testData cleanup, other essential Refactors
Abstracted out NASDetails with formating and expansion and all that to the NASDetails.jsx. Instead of api fetch both departure and destination on NAS it now fetches either airport or (departure and destination). Removed searchComponent from the weatherCard, flightCard and GateCard and kept a singular one in details.jsx for showing on all components. Cleaned up TabFormat to use only flightData, weather, and NAS as variables and those three variables contain essential details and they are exploded within TabFormat.